+++ kotest-framework/kotest-framework-datatest/src/jvmTest/kotlin/io/kotest/datatest/ContainerDataTestNameFunctionTest.kt
@@ -6,6 +6,7 @@ import io.kotest.core.spec.style.FunSpec
 import io.kotest.core.test.TestCase
 import io.kotest.core.test.TestResult
 import io.kotest.matchers.shouldBe
+import kotlinx.coroutines.delay
 
 @ExperimentalKotest
 class ContainerDataTestNameFunctionTest : FunSpec({
